🦷 🔄 Extract inversion utilities
This serves a stopgap to reduce the diff in #752
