+++ src/pykeen/triples/__init__.py
@@ -3,15 +3,7 @@
 """Classes for creating and storing training data from triples."""
 
 from .instances import Instances, LCWAInstances, SLCWAInstances
-from .triples_factory import (
-    AnyTriples,
-    CoreTriplesFactory,
-    KGInfo,
-    RelationInverter,
-    TriplesFactory,
-    get_mapped_triples,
-    relation_inverter,
-)
+from .triples_factory import AnyTriples, CoreTriplesFactory, KGInfo, TriplesFactory, get_mapped_triples
 from .triples_numeric_literals_factory import TriplesNumericLiteralsFactory
 
 __all__ = [
@@ -20,8 +12,6 @@ __all__ = [
     "SLCWAInstances",
     "KGInfo",
     "CoreTriplesFactory",
-    "RelationInverter",
-    "relation_inverter",
     "TriplesFactory",
     "TriplesNumericLiteralsFactory",
     "get_mapped_triples",
