+++ src/pykeen/triples/triples_factory.py
@@ -7,7 +7,6 @@ import logging
 import pathlib
 import re
 import warnings
-from abc import abstractmethod
 from typing import (
     Any,
     Callable,
@@ -23,7 +22,6 @@ from typing import (
     Set,
     TextIO,
     Tuple,
-    TypeVar,
     Union,
     cast,
 )
@@ -37,6 +35,7 @@ from .instances import BatchedSLCWAInstances, LCWAInstances, SubGraphSLCWAInstan
 from .splitting import split
 from .utils import TRIPLES_DF_COLUMNS, load_triples, tensor_to_df
 from ..constants import COLUMN_LABELS
+from ..inverse import relation_inverter_resolver
 from ..typing import EntityMapping, LabeledTriples, MappedTriples, RelationMapping, TorchRandomHint
 from ..utils import (
     ExtraReprMixin,
@@ -58,8 +57,6 @@ __all__ = [
     "cat_triples",
     "splits_steps",
     "splits_similarity",
-    "RelationInverter",
-    "relation_inverter",
     "AnyTriples",
     "get_mapped_triples",
 ]
@@ -175,59 +172,6 @@ def _ensure_ids(
 ) -> Collection[int]:
     """Convert labels to IDs."""
     return [label_to_id[l_or_i] if isinstance(l_or_i, str) else l_or_i for l_or_i in labels_or_ids]
-
-
-RelationID = TypeVar("RelationID", int, torch.LongTensor)
-
-
-class RelationInverter:
-    """An interface for inverse-relation ID mapping."""
-
-    # TODO: method is_inverse?
-
-    @abstractmethod
-    def get_inverse_id(self, relation_id: RelationID) -> RelationID:
-        """Get the inverse ID for a given relation."""
-        # TODO: inverse of inverse?
-        raise NotImplementedError
-
-    @abstractmethod
-    def _map(self, batch: torch.LongTensor, index: int = 1) -> torch.LongTensor:
-        raise NotImplementedError
-
-    @abstractmethod
-    def invert_(self, batch: torch.LongTensor, index: int = 1) -> torch.LongTensor:
-        """Invert relations in a batch (in-place)."""
-        raise NotImplementedError
-
-    def map(self, batch: torch.LongTensor, index: int = 1, invert: bool = False) -> torch.LongTensor:
-        """Map relations of batch, optionally also inverting them."""
-        batch = self._map(batch=batch, index=index)
-        return self.invert_(batch=batch, index=index) if invert else batch
-
-
-class DefaultRelationInverter(RelationInverter):
-    """Maps normal relations to even IDs, and the corresponding inverse to the next odd ID."""
-
-    # docstr-coverage: inherited
-    def get_inverse_id(self, relation_id: RelationID) -> RelationID:  # noqa: D102
-        return relation_id + 1
-
-    # docstr-coverage: inherited
-    def _map(self, batch: torch.LongTensor, index: int = 1, invert: bool = False) -> torch.LongTensor:  # noqa: D102
-        batch = batch.clone()
-        batch[:, index] *= 2
-        return batch
-
-    # docstr-coverage: inherited
-    def invert_(self, batch: torch.LongTensor, index: int = 1) -> torch.LongTensor:  # noqa: D102
-        # The number of relations stored in the triples factory includes the number of inverse relations
-        # Id of inverse relation: relation + 1
-        batch[:, index] += 1
-        return batch
-
-
-relation_inverter = DefaultRelationInverter()
 
 
 @dataclasses.dataclass
@@ -423,6 +367,7 @@ class CoreTriplesFactory(KGInfo):
         if metadata is None:
             metadata = dict()
         self.metadata = metadata
+        self.relation_inverter = relation_inverter_resolver.make(query=None)
 
     @classmethod
     def create(
@@ -514,7 +459,7 @@ class CoreTriplesFactory(KGInfo):
         """Get the inverse relation identifier for the given relation."""
         if not self.create_inverse_triples:
             raise ValueError("Can not get inverse triple, they have not been created.")
-        return relation_inverter.get_inverse_id(relation_id=relation)
+        return self.relation_inverter.get_inverse_id(relation_id=relation)
 
     def _add_inverse_triples_if_necessary(self, mapped_triples: MappedTriples) -> MappedTriples:
         """Add inverse triples if they shall be created."""
@@ -524,8 +469,8 @@ class CoreTriplesFactory(KGInfo):
         logger.info("Creating inverse triples.")
         return torch.cat(
             [
-                relation_inverter.map(batch=mapped_triples),
-                relation_inverter.map(batch=mapped_triples, invert=True).flip(1),
+                self.relation_inverter.map(batch=mapped_triples),
+                self.relation_inverter.map(batch=mapped_triples, invert=True).flip(1),
             ]
         )
